More sensible order of commands, and short-term memory within work

+++ src/Application/Common/Configuration/AppSettings.cs
@@ -2,6 +2,6 @@ namespace Tarik.Application.Common;
 
 public class AppSettings
 {
-    public int? WorkItemPollingIntervalInMinutes { get; set; }
+    public float? WorkItemPollingIntervalInMinutes { get; set; }
     public string? GitHubPAT { get; set; }
 }
